Support rollback!
Ahhh aH this feels really good to have working! Have one key bug to fix which causes
issues, but otherwise... damn it works! ^_^

Three key parts of commit to get rollback working finally:
1. Support storing and restoring snapshot data
2. Predict inputs if haven't received remote player's input
3. When finally receiving remote player's input, compare to predictions. If predictions
off, then simply rollback (which is pretty simple logic but placed in the GameManager
which isn't part of this repo)