Minor: Add helper function to TransformComponent to get forward direction
This represents a shift in thinking- wanted to avoid adding helper functions to
components to keep them lean, but this forward direction *is* represented by the
component. Thus, no reason to not add the helper function directly to the struct itself.

Definitely way cleaner than having a mess of helper functions in arbitrary "entity"
classes ^_^