Fix final failing box-capsule base collision unit test!
Wow, this was a surprising amount of work! Really glad that eventually got to a
solution pretty happy with aH!

Also very happy that came back to raycast and re-commented it out with a full
understanding of what the heck is going on (this time around). Both passage of time
and me working on various other game physics problems really helped a ton for
building a decent foundation for these sorts of problems.

Still need to update other linetest and raycast methods to consider touching of
surface to NOT be an intersection, but I can likely deal with that as the need arises