Fix linetest with capsule!
Yay for finally having energy to figure this out while trying to sleep!
Who needs to sleep on time to follow a neat schedule made for next day =')

Seriously though, really glad had the energy to come back and PROPERLY fix this.
Unfortunately, it seems that this hasn't actually fixed the box-capsule collision being
very broken, so more work (as expected)!