Rename conflicting param name in Collider
Surprisingly failing to compile with UE5 (C++20) even though the overlapping
variable names are in two different methods.

Quick fix, but this also depicts how it may be worth adding the "m" prefix to the
member variables here as well. Really need to finally choose and commit to a style
guide...