Add MoveDataProvider to SimContext... and realize SimContext needs to be removed from this project
Yay repo commit process forcing me to realize this file is actually in the core project- when an include is
actually in the SimLayer project.

Barely getting away with this due to lack of unit tests and former project include setup in actual game.
Will refactor/rework in future... eventually!