Add additional methods to input buffer retrieval for code flexibility, and some other minor misc changes
New input buffer related changes are a mess, but following philosophy of getting things done at all
rather than (never) perfectly.

Almost certainly going to come back to this input buffer code and rewrite it, cuz it definitely bugs me
how messy it is. However, do want to use it more in order to see what would be best ways to tweak it